services/model-service/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
email-validator==2.1.0
requests==2.31.0
pyyaml==6.0.1

opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.24.0
numpy==1.26.4
pandas==2.1.4

torch==2.2.2
torchvision==0.17.2
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3; sys_platform != 'darwin'

boto3==1.34.0
botocore==1.34.0
